Add TTS notifications to Wear OS
